refactor: 아톰 메인 index.ts 파일 추가 및 각 컴포넌트에서 아톰 import 경로 수정
