+++ front/config/webpack.common.js
@@ -19,7 +19,7 @@ export default {
 			'@assets': path.resolve('assets'),
 			'@pages': path.resolve('src', 'pages'),
 			'@tools': path.resolve('src', 'tools'),
-			'@recoil': path.resolve('src', 'recoil'),
+			'@recoil': path.resolve('src', 'recoil', 'index'),
 			'@lib': path.resolve('src', 'lib'),
 		},
 	},
