+++ front/src/pages/balance/Balance.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import { Redirect } from 'react-router-dom';
 import { useRecoilValue } from 'recoil';
 import { BALANCE_TYPE, STATUS_TYPE, IUser, IHistory } from '@src/types';
-import userAtom from '@recoil/user';
+import { userAtom } from '@recoil';
 import toDateString from '@src/common/utils/toDateString';
 import { getBalance } from '@lib/api';
 import { NINE_HOURS_IN_MILLISECONDS, ONE_MONTH_IN_MILLISECONDS } from '@common/constants';
