+++ front/src/pages/trade/bidAsk/BidAsk.tsx
@@ -2,8 +2,7 @@ import React, { useEffect, useState } from 'react';
 import TOAST from '@lib/toastify';
 import { useRecoilState, useRecoilValue } from 'recoil';
 import { IUser, IHoldStockItem } from '@src/types';
-import bidAskPriceAtom from '@recoil/bidAskPrice';
-import userAtom from '@recoil/user';
+import { bidAskPriceAtom, userAtom } from '@recoil';
 import { getUserAskAvailable, getUserBidAvailable } from '@common/utils/getAvailableAmount';
 import Emitter from '@common/utils/eventEmitter';
 import BidAskType from './BidAskType';
