+++ front/src/pages/trade/chart/candle/CandleBackground.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useRef } from 'react';
 import { useRecoilValue } from 'recoil';
 import { IChartItem, Theme, IUser } from '@src/types';
-import userAtom from '@recoil/user';
+import { userAtom } from '@recoil';
 import { getMaxValue, getMinValue, getTextColor, getLegendColor, getText } from '../common';
 
 interface IProps {
