+++ front/src/pages/trade/chart/volume/VolumeLegend.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useRef } from 'react';
 import { useRecoilValue } from 'recoil';
 import { IChartItem, Theme, IUser } from '@src/types';
-import userAtom from '@recoil/user';
+import { userAtom } from '@recoil';
 import {
 	ICrossLine,
 	VOLUME_CANVAS_TOP_BOT_PADDING as CANVAS_PADDING,
